+++ lib/langtool/examples/example.ex
@@ -6,8 +6,6 @@ defmodule Langtool.Examples.Example do
   schema "examples" do
     belongs_to :sentence, Sentence
     belongs_to :translation, Translation
-
-    timestamps()
   end
 
   @doc false