+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - plug to check email confirmation
 - Translations controller, templates
 - render/edit existed translations
+- create new translations
 
 ## [0.2.1] - 2019-01-06
 ### Added